move dgetc2 to lapack
