+++ blas/gonum/level2float64_test.go
@@ -18,10 +18,6 @@ func TestDger(t *testing.T) {
 	testblas.DgerTest(t, impl)
 }
 
-func TestDgetc2(t *testing.T) {
-	testblas.Dgetc2Test(t, impl)
-}
-
 func TestDtxmv(t *testing.T) {
 	testblas.DtxmvTest(t, impl)
 }
