+++ lapack/testlapack/dgetc2.go
@@ -1,15 +1,17 @@
-// Copyright ©2021 The Gonum Authors. All rights reserved.
+// Copyright ©2015 The Gonum Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package testblas
+package testlapack
 
-import "testing"
+import (
+	"testing"
+)
 
 type Dgetc2er interface {
 	Dgetc2(n int, a []float64, lda int, ipiv, jpiv []int) (k int)
 }
 
-func Dgetc2Test(t *testing.T, blasser Dgetc2er) {
+func Dgetc2Test(t *testing.T, impl Dgetc2er) {
 
 }
